+++ src/features/layout.ts
@@ -3,7 +3,7 @@ import type { FeatureConfig } from "../types";
 import { scriptInfo } from "../constants";
 import { error, getResourceUrl, log, warn } from "../utils";
 import { SiteEventsMap, siteEvents } from "../events";
-import { openMenu } from "./menu/menu_old";
+import { openMenu } from "../menu/menu_old";
 import { getGeniusUrl, createLyricsBtn, sanitizeArtists, sanitizeSong, getLyricsCacheEntry, splitVideoTitle } from "./lyrics";
 import { featInfo } from "./index";
 import "./layout.css";
